Separate firmware updater as a class, the refactoring is only partially done
